Use size_t for sv size (-Wconversion)
rasqal_rowsource_sv_context field variables_count changes to size_t
to capture full sv size

+++ src/rasqal_format_sv.c
@@ -299,7 +299,7 @@ typedef struct
 
   /* Variables table allocated for variables in the result set */
   rasqal_variables_table* vars_table;
-  int variables_count;
+  size_t variables_count;
 
   unsigned int flags;
